requirements.txt
python-telegram-bot
python-dotenv
google-generativeai
aiosqlite
apscheduler
httpx==0.27.0
solana==0.32.0
websockets==12.0